+++ src/assets/js/i18n/locales/en.js
@@ -223,6 +223,7 @@ export default {
   "matrix.dest.oscilloscope.Y": "Scope Y",
   "matrix.dest.oscilloscope.X": "Scope X",
   "matrix.dest.oscFreqCV": "Osc {osc} Freq CV",
+  "matrix.dest.oscSync": "Osc {osc} Sync",
   "matrix.dest.outputLevelCV": "Out {bus} Level CV",
   "matrix.tooltip.format": "{source} → {dest}",
   "matrix.tooltip.unknownSource": "Row {row}",